Big refactoring, pulling lots of files further up.
A lot has been moved in this commit. Mainly, this stems from the fact that `ts/mod.rs` and `ts/transition_system` have been
fused into a single `src/transition_system.rs` which has brought a lot of modules further up
We should now continue sorting through the `transition_system` module. There should be less top-level files in that directory.
After that the major things of refactoring are done. But mathematical stuff should be moved to a `theory` module or something
similar, see also #17 .
