+++ src/transition_system/operations.rs
